pyproject.toml
[project]
name = "airbnb-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "apache-airflow>=3.1.8",
    "dbt-core>=1.11.7",
    "dbt-snowflake>=1.11.3",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "seaborn>=0.13.2",
    "sweetviz>=2.3.1",
]
